requirement.txt
autopep8==1.6.0
Pillow==9.0.1
pycodestyle==2.8.0
toml==0.10.2